% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{beta_P_value}
\alias{beta_P_value}
\title{Global test of beta for moderate-dimensional inferences using Rcpp}
\usage{
beta_P_value(Y, X, beta)
}
\arguments{
\item{Y}{a numeric vector}

\item{X}{a numeric matrix}

\item{beta}{the null beta}
}
\value{
P value of beta global test in moderate-dimensional
}
\description{
Global test of beta for moderate-dimensional inferences using Rcpp
}
\examples{
\dontrun{
n=1000
pn=900
beta<-runif(pn)
sigma<-runif(n)
X<-matrix(rnorm(pn*n),nrow = n,ncol = pn)
epsilon<-numeric(n)
for(j in 1:n)epsilon[j]<-rnorm(1,mean = 0,sd = sigma)
Y<-X\%*\%beta+epsilon
beta_P_value(Y,X,rep(0,pn))
beta_P_value(Y,X,beta)
}
}
