% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CommunityC}
\alias{CommunityC}
\title{CommunityC}
\usage{
CommunityC(embedding_mat, k, iter, s)
}
\arguments{
\item{embedding_mat}{the data embedding matrix}

\item{k}{the number of communities}

\item{iter}{the number of iterations}

\item{s}{the list of node name}
}
\value{
result the list of node name which in the same community
}
\description{
Community Detection by k-means clustering using k and embedding matrix obtained in RS_PA
}
\examples{
\dontrun{
set.seed(0)
n <- 100
p <- 3
embedding_mat <- matrix(rnorm(n * p), nrow = n, ncol = p)
s <- paste0("node", 1:n)
k <- 3
iter <- 50
CommunityC(embedding_mat, k, iter, s)
data(dolphins)
dolphins_mat=get.adjacency(dolphins,sparse = FALSE)
s=get.vertex.attribute(dolphins)$label
k=RS_PA(dolphins_mat,300,0.95)$k
embedding_mat=RS_PA(dolphins_mat,300,0.95)$embedding_mat
iter <- 50
CommunityC(embedding_mat, k, iter, s)
}
}
