% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp.R
\name{dolphins}
\alias{dolphins}
\title{Dolphins dataset}
\description{
This dataset contains an undirected social network of frequent associations between 62 dolphins in a community living off Doubtful Sound, New Zealand, as compiled by Lusseau et al. (2003).
}
\examples{
\dontrun{
data(dolphins)
RS_PA(dolphins)
}
}
