% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp.R
\name{football}
\alias{football}
\title{Football dataset}
\description{
This dataset contains the network of American football games between Division IA colleges during regular season Fall 2000, as compiled by M. Girvan and M. Newman.
}
\examples{
\dontrun{
data(football)
RS_PA(football)
}
}
