% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp.R
\name{karate}
\alias{karate}
\title{Karate dataset}
\description{
This dataset contains the network of friendships between the 34 members of a karate club at a US university, as described by Wayne Zachary in 1977.
}
\examples{
\dontrun{
data(karate)
RS_PA(karate)
}
}
