% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imp.R
\name{polbooks}
\alias{polbooks}
\title{Books about US politics dataset}
\description{
Nodes represent books about US politics sold by the online bookseller Amazon.com.  Edges represent frequent co-purchasing of books by the same buyers, as indicated by the "customers who bought this book also bought these other books" feature on Amazon.
}
\examples{
\dontrun{
data(polbooks)
RS_PA(polbooks)
}
}
