\name{SA23204166-package}
\alias{SA23204166-package}
\docType{package}
\title{
  The Statistical Computing project of Jin Yicheng SA23204166.
}
\description{
  R and Rcpp function are included in this package for the Statistical Computing project of Jin Yicheng SA23204166.
}
\author{
Jin Yicheng 
Maintainer: Jin Yicheng <jyc2002@mail.ustc.edu.cn>
}
\keyword{ package }

