% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RS_PA.R
\name{RS_PA}
\alias{RS_PA}
\title{Random Signflips parallel analysis using R}
\usage{
RS_PA(X, T, a)
}
\arguments{
\item{X}{the sample matrix}

\item{T}{the number of trials}

\item{a}{the percentile}
}
\value{
k the number of leading singular values above the 𝛼-percentile of their signflipped analogues

embedding_mat the embedding matrix of nodes with community k by PCA
}
\description{
Random Signflips parallel analysis using R
}
\examples{
\dontrun{
    mat=matrix(rbinom(50^2,1,0.2),ncol=50)
    upper_triangle <- upper.tri(mat)
    sim_data <- matrix(0, nrow = nrow(mat), ncol = ncol(mat))
    sim_data[upper_triangle] <- mat[upper_triangle]
    sim_data[lower.tri(sim_data)] <- t(sim_data)[lower.tri(sim_data)]
    RS_PA(sim_data,300,0.95)
    data(dolphins)
    dolphins_mat=get.adjacency(dolphins,sparse = FALSE)
    RS_PA(dolphins_mat,300,0.95)
    data(football)
    football_mat=get.adjacency(football,sparse = FALSE)
    RS_PA(football_mat,300,0.95)
    data(karate)
    karate_mat=get.adjacency(karate,sparse = FALSE)
    RS_PA(karate_mat,300,0.95)
    data(polbooks)
    polbooks_mat=get.adjacency(polbooks,sparse = FALSE)
    RS_PA(polbooks_mat,300,0.95)
}
}
